Relocate misplaced unit tests.
